+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
  - Set `WINEDLLPATH` and `WINELOADER` environment variables.
+ - Add a workaround for a Zenity bug that causes the GUI to crash when certain versions of Zenity are used.
 
 ## [1.2.1] - 2019-04-08
 ### Changed